Add frontend owner assertion
Added AssertOwner that compares the addressed used on the pixel
stored on the blockchain. This function is supposed to run after the user has made the transaction and changed the pixel owner on the Blockchain.
If this function returns true, the user is now allowed to change the pixel color.